fix(ui): Remove circular dependencies
