+++ apps/ui/src/core/store/tests/useWalletAdapter.test.ts
@@ -1,13 +1,14 @@
 import { PublicKey } from "@solana/web3.js";
 import { act, renderHook } from "@testing-library/react-hooks";
 
-import { useNotification as notificationStore, useWalletAdapter } from "..";
 import { Protocol } from "../../../config";
 import type { WalletAdapter } from "../../../models";
 import { WalletServiceId } from "../../../models";
 import { EvmWeb3WalletAdapter } from "../../../models/wallets/adapters/evm";
 import { PhantomAdapter } from "../../../models/wallets/adapters/solana/PhantomAdapter";
+import { useNotification as notificationStore } from "../useNotification";
 import type { WalletAdapterState } from "../useWalletAdapter";
+import { useWalletAdapter } from "../useWalletAdapter";
 
 const getProtocolAdapter = (state: WalletAdapterState, protocol: Protocol) =>
   protocol === Protocol.Evm ? state.evm : state.solana;
