+++ apps/ui/src/hooks/evm/useEvmWallet.ts
@@ -1,12 +1,6 @@
 import { selectEvmAdapter } from "../../core/selectors";
 import { useWalletAdapter } from "../../core/store";
-import type { EvmWalletAdapter } from "../../models";
-
-export interface EvmWalletInterface {
-  readonly wallet: EvmWalletAdapter | null;
-  readonly address: string | null;
-  readonly connected: boolean;
-}
+import type { EvmWalletInterface } from "../../models";
 
 export const useEvmWallet = (): EvmWalletInterface => {
   const wallet = useWalletAdapter(selectEvmAdapter);
