+++ apps/ui/src/hooks/interaction/usePrepareSplTokenAccountMutation.test.ts
@@ -2,14 +2,14 @@ import type { AccountInfo as TokenAccount } from "@solana/spl-token";
 import { act, renderHook } from "@testing-library/react-hooks";
 import { useQueryClient } from "react-query";
 
-import type { SolanaWalletInterface } from "..";
-import { useSolanaConnection, useSolanaWallet } from "..";
 import { selectGetInteractionState } from "../../core/selectors";
 import { useInteractionState } from "../../core/store";
 import { MOCK_SOL_WALLET } from "../../fixtures";
 import { MOCK_INTERACTION_STATE } from "../../fixtures/swim/interactionState";
+import type { SolanaWalletInterface } from "../../models";
 import { createSplTokenAccount } from "../../models";
 import { mockOf, renderHookWithAppContext } from "../../testUtils";
+import { useSolanaConnection, useSolanaWallet } from "../solana";
 
 import { usePrepareSplTokenAccountMutation } from "./usePrepareSplTokenAccountMutation";
