+++ apps/ui/src/hooks/solana/useSolanaWallet.ts
@@ -1,12 +1,6 @@
 import { selectSolanaAdapter } from "../../core/selectors";
 import { useWalletAdapter } from "../../core/store";
-import type { SolanaWalletAdapter } from "../../models";
-
-export interface SolanaWalletInterface {
-  readonly wallet: SolanaWalletAdapter | null;
-  readonly address: string | null;
-  readonly connected: boolean;
-}
+import type { SolanaWalletInterface } from "../../models";
 
 export const useSolanaWallet = (): SolanaWalletInterface => {
   const wallet = useWalletAdapter(selectSolanaAdapter);
